requirements.txt
astral==2.2
certifi==2021.10.8
charset-normalizer==2.0.9
colorzero==2.0
gpiozero==1.6.2
greenlet==1.1.2
idna==3.3
numpy==1.21.4
pandas==1.3.5
psycopg2==2.9.2
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
RPi.GPIO==0.7.0
six==1.16.0
SQLAlchemy==1.4.28
urllib3==1.26.7
